backend/requirements.txt
jellyfish-server-sdk @ git+https://github.com/jellyfish-dev/python-server-sdk.git@fix-events
